pyproject.toml
[tool.poetry]
name = "hacker-news-small-sites-search"
version = "0.1.0"
description = ""
authors = ["Alex Wendland <me@alexwendland.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
lxml = "^5.1.0"
llama-index-core = "^0.10.12"
llama-index-embeddings-openai = "^0.1.6"
chromadb = "^0.4.23"
html2text = "^2024.2.25"
langchain-community = "^0.0.24"
sentence-transformers = "^2.4.0"
fastui = "^0.5.2"
fastapi = "^0.110.0"
python-multipart = "^0.0.9"
jinja2 = "^3.1.3"
cachetools = "^5.3.3"
feedparser = "^6.0.11"


[tool.poetry.group.dev.dependencies]
pyright = "^1.1.352"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
